Abstract mkdirs with options
