+++ underfs/s3/src/main/java/alluxio/underfs/s3/S3UnderFileSystem.java
@@ -247,38 +247,6 @@ public final class S3UnderFileSystem extends ObjectUnderFileSystem {
   }
 
   @Override
-  public boolean mkdirs(String path, MkdirsOptions options) throws IOException {
-    if (path == null) {
-      return false;
-    }
-    if (isDirectory(path)) {
-      return true;
-    }
-    if (isFile(path)) {
-      LOG.error("Cannot create directory {} because it is already a file.", path);
-      return false;
-    }
-    if (!options.getCreateParent()) {
-      if (parentExists(path)) {
-        // Parent directory exists
-        return mkdirsInternal(path);
-      } else {
-        LOG.error("Cannot create directory {} because parent does not exist", path);
-        return false;
-      }
-    }
-    // Parent directories should be created
-    if (parentExists(path)) {
-      // Parent directory exists
-      return mkdirsInternal(path);
-    } else {
-      String parentKey = getParentKey(path);
-      // Recursively make the parent folders
-      return mkdirs(parentKey, true) && mkdirsInternal(path);
-    }
-  }
-
-  @Override
   public InputStream open(String path) throws IOException {
     try {
       path = stripPrefixIfPresent(path);
